artdb/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
astroid==2.0
brotli==1.0.7             # via whitenoise
certifi==2018.8.24        # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
dal-admin-filters==0.3.3
django-appconf==1.0.2
django-autocomplete-light==3.3.2
django-cas-ng==3.5.10
django-cleanup==2.1.0
django-debug-toolbar==1.10.1
django-environ==0.4.5
django-imagekit==4.0.2
django-js-asset==1.1.0    # via django-mptt
django-mass-edit==3.2.0
django-mptt==0.9.1
django-ordered-model==3.1.1
django-redis==4.9.0
django-rest-framework==0.1.0
django-versatileimagefield==1.10
django==2.2.1
djangorestframework==3.8.2  # via django-rest-framework
gunicorn==19.9.0
idna==2.7                 # via requests
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.5               # via python-cas, python-pptx
mccabe==0.6.1
pilkit==2.0
pillow==5.1.0
pip-tools==3.1.0
psutil==5.4.7             # via rainbow-saddle
psycopg2-binary==2.7.5
pylint==2.0.0
python-cas==1.4.0         # via django-cas-ng
python-pptx==0.6.16
pytz==2018.4
rainbow-saddle==0.4.0
redis==2.10.6             # via django-redis
requests==2.19.1
six==1.11.0
sqlparse==0.2.4           # via django, django-debug-toolbar
typed-ast==1.1.0
urllib3==1.23             # via requests
whitenoise[brotli]==4.1.2
wrapt==1.10.11
xlsxwriter==1.1.2         # via python-pptx
